Address @JaKXz findings.
